+++ mediation/mediatedviews/InMobi/src/com/appnexus/opensdk/mediatednativead/InMobiSettings.java
@@ -23,6 +23,7 @@ public class InMobiSettings {
     static String KEY_LANDING_URL = "landingURL";
     static String KEY_DESCRIPTION = "description";
     public static String INMOBI_APP_ID = "";
+    public static String NATIVE_ELEMENT_OBJECT = "element";
 
     public static void setInMobiAppId(String key, Context context) {
         INMOBI_APP_ID = key;